requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
backports.zoneinfo==0.2.1
certifi==2022.9.24
charset-normalizer==2.1.1
coverage==6.4.4
deepdiff==6.2.1
Deprecated==1.2.13
discord-webhook==0.17.0
discord.py==2.0.1
Django==4.1.2
django-cors-headers==3.13.0
django-debug-toolbar==3.7.0
django-query-profiler==0.9
django-redis==5.2.0
djangorestframework==3.13.1
frozenlist==1.3.3
idna==3.4
isort==5.10.1
mmh3==3.0.0
mo-dots==9.230.22310
mo-future==6.230.22310
mo-imports==7.230.22310
mo-parsing==8.233.22310
mo-sql-parsing==8.235.22313
multidict==6.0.2
ordered-set==4.1.0
packaging==21.3
pyparsing==3.0.9
python-dotenv==0.21.0
pytz==2022.2.1
redis==4.3.4
requests==2.28.1
sqlparse==0.4.2
tzdata==2022.2
urllib3==1.26.12
wrapt==1.14.1
yarl==1.8.1
